requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
sqlalchemy==2.0.25
bcrypt==4.1.2
python-doctr[torch]==0.7.0
pytesseract==0.3.10
PyMuPDF==1.23.8
google-generativeai==0.8.3
weasyprint==61.2
matplotlib>=3.8.0
numpy>=1.26.0
pydyf==0.10.0
python-dotenv==1.0.0
itsdangerous==2.1.2
python-magic==0.4.27
pydantic-settings
psycopg2-binary==2.9.9
resend==2.0.0
playwright==1.40.0
extract-msg==0.48.7
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
pillow==10.2.0
stripe==7.0.0
xhtml2pdf==0.2.14
psutil==5.9.0
slowapi==0.1.9

